Check SCM URLs before checking project URL
The project URL may be a vanity domain that could possibly
return a false positive if it matches the repository id
heuristic and therefore the SCM URLs are more likely to
be accurate if referencing a GitHub repository and so they
should be checked first.

The project URL is still checked but is now last in the
order.

Closed issue #11