Ensure .nojekyll file is always at the root of the tree
If a custom path was specified and the noJekyll option was
enabled then the .nojekyll file would be under the custom
path but it is only valid if at the root of the repository.

No longer create a local .nojekyll file, instead just
add a tree entry with a .nojekyll path and an empty blob
when the option is enabled.